requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform has been tested: win-64 linux-x86_86

requests~=2.28.2
argparse~=1.4.0
gmssl~=3.2.2
mmh3~=5.1.0
Pillow~=10.1.0
crypto~=1.4.1